requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
openai
httpx
requests
gradio
plotly
pandas
pyserial
pytest
python-multipart
PyQt6
PyQt6-Charts
pyqtgraph
PyOpenGL
pyinstaller
